[#17] Enhancement hide sidebar
- return `null` instead of `Sidebar` when no user is present
- replace pro icon `fa fa-signout` with button and text
- fix undefined router wrapping Sideabar using `withRouter` HoC
- cleanup the return
- fix some console warnings

Closes https://github.com/CityOfZion/neochat/issues/17
